extract game mechacnis to GameHandler class
